refactor(listener): extract memory list command (#2073)
